requirements.txt
PyGithub>=2.1.1
gitpython>=3.1.40
requests>=2.31.0
python-dotenv>=1.0.0
click>=8.1.7
rich>=13.7.0
tabulate>=0.9.0
